updated login and refersh login features to issue expiry duration instead of date time

+++ Flipcart-Clone-API/src/main/java/com/self/flipcart/util/CookieManager.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class CookieManager {
 
-    public String configure(String name,String value, int maxAge) {
+    public String configure(String name,String value, long maxAge) {
         ResponseCookie cookie = ResponseCookie.from(name, value)
                 .domain("localhost")
                 .path("/")
